Use open weather API for fetching weather data for the current user location

+++ Weather/Weather/features/weather detail/domain/Weather.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct Weather {
-    var temperature: Float
+    var temperature: Double
     var temperatureUnit: UnitTemperature
     var geoCoordinates: GeoCoordinates
     var locationName: String?
